Initial commit
This plugin is designed to create a single document per player,
Addressable under domain "bsinv", key "inventory". This document should
only be accessed via the web api, and has a structure similar to the NBT
format used to store inventories. Some nbt data is missing (notably
firework metadata among others).

The aim is to test and release this alongside the beta of 0.8 with a
suitable web api to display this information, as a showcase of how to
use documents.